bot/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.1
python-dotenv==1.0.1
SQLAlchemy==2.0.34
alembic==1.13.2
apscheduler==3.10.4
httpx==0.27.2
numpy==2.1.1
email-validator==2.2.0
itsdangerous==2.2.0
pytest==8.3.2
pytest-asyncio==0.23.8
